feat!: remove legacy architecture (Paper) support
The library is now New Architecture (Fabric/TurboModules) only.

Android:
- Delete the src/oldarch source set (Paper ViewGroupManager and
  ReactContextBaseJavaModule wrappers).
- Merge the arch-specific wrappers and the shared *Impl classes into
  single classes in src/main: RNCWebViewManagerImpl.kt + the Fabric
  manager become RNCWebViewManager.kt, and RNCWebViewModuleImpl.java +
  the TurboModule wrapper become RNCWebViewModule.java.
- Always apply the com.facebook.react plugin, drop the conditional
  source sets and the IS_NEW_ARCHITECTURE_ENABLED build config field;
  fail the build if newArchEnabled=false is set explicitly.
- The loadUrl command now resets the progress-changed filter, which was
  previously only done on the Paper receiveCommand path.

iOS/macOS/visionOS:
- Delete the Paper RNCWebViewManager and the RCTConvert
  WKDataDetectorTypes category it used.
- Remove all RCT_NEW_ARCH_ENABLED guards; RNCWebViewImpl cleanup now
  always happens via destroyWebView (Fabric prepareForRecycle) instead
  of removeFromSuperview.
- Simplify the podspec to always use install_modules_dependencies.
- Opt the macOS and visionOS examples into the new architecture:
  react-native-test-app only forces it for react-native >= 0.82, and
  react-native-macos 0.81 / react-native-visionos 0.79 still default
  the example to the legacy architecture otherwise.

JS:
- Stop passing the legacy `source` prop to the native component and
  drop the unused Paper macOS component wrapper; macOS uses the shared
  codegen component.
- Make messagingModuleName and onLoadingSubResourceError optional in
  the codegen spec (Android-only, other platforms never provided them;
  previously masked by a ts-expect-error on the legacy source prop).
- macOS now stringifies injectedJavaScriptObject like iOS/Android.

Docs: update the READMEs and Getting Started guide to state that the
new architecture is required.

BREAKING CHANGE: the legacy (Paper) React Native architecture is no
longer supported; the New Architecture is required.
