--- android/src/newarch/com/reactnativecommunity/webview/RNCWebViewModule.java
@@ -1,47 +1,0 @@
-package com.reactnativecommunity.webview;
-
-import android.app.DownloadManager;
-import android.net.Uri;
-import android.webkit.ValueCallback;
-
-import androidx.annotation.NonNull;
-
-import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.module.annotations.ReactModule;
-
-@ReactModule(name = RNCWebViewModuleImpl.NAME)
-public class RNCWebViewModule extends NativeRNCWebViewModuleSpec {
-    final private RNCWebViewModuleImpl mRNCWebViewModuleImpl;
-
-    public RNCWebViewModule(ReactApplicationContext reactContext) {
-        super(reactContext);
-        mRNCWebViewModuleImpl = new RNCWebViewModuleImpl(reactContext);
-    }
-
-    @Override
-    public void shouldStartLoadWithLockIdentifier(boolean shouldStart, double lockIdentifier) {
-        mRNCWebViewModuleImpl.shouldStartLoadWithLockIdentifier(shouldStart, lockIdentifier);
-    }
-
-    public boolean startPhotoPickerIntent(final ValueCallback<Uri[]> callback, final String[] acceptTypes, final boolean allowMultiple, final boolean isCaptureEnabled) {
-        return mRNCWebViewModuleImpl.startPhotoPickerIntent(acceptTypes, allowMultiple, callback, isCaptureEnabled);
-    }
-
-    public void setDownloadRequest(DownloadManager.Request request) {
-        mRNCWebViewModuleImpl.setDownloadRequest(request);
-    }
-
-    public void downloadFile(String downloadingMessage) {
-        mRNCWebViewModuleImpl.downloadFile(downloadingMessage);
-    }
-
-    public boolean grantFileDownloaderPermissions(String downloadingMessage, String lackPermissionToDownloadMessage) {
-        return mRNCWebViewModuleImpl.grantFileDownloaderPermissions(downloadingMessage, lackPermissionToDownloadMessage);
-    }
-
-    @NonNull
-    @Override
-    public String getName() {
-        return RNCWebViewModuleImpl.NAME;
-    }
-}
