--- android/src/oldarch/com/reactnativecommunity/webview/RNCWebViewManager.java
@@ -1,335 +1,0 @@
-package com.reactnativecommunity.webview;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-
-import com.facebook.react.bridge.ReadableArray;
-import com.facebook.react.bridge.ReadableMap;
-import com.facebook.react.common.MapBuilder;
-import com.facebook.react.uimanager.ThemedReactContext;
-import com.facebook.react.uimanager.ViewGroupManager;
-import com.facebook.react.uimanager.annotations.ReactProp;
-import com.facebook.react.views.scroll.ScrollEventType;
-import com.reactnativecommunity.webview.events.TopCustomMenuSelectionEvent;
-import com.reactnativecommunity.webview.events.SubResourceErrorEvent;
-import com.reactnativecommunity.webview.events.TopHttpErrorEvent;
-import com.reactnativecommunity.webview.events.TopLoadingErrorEvent;
-import com.reactnativecommunity.webview.events.TopLoadingFinishEvent;
-import com.reactnativecommunity.webview.events.TopLoadingProgressEvent;
-import com.reactnativecommunity.webview.events.TopLoadingStartEvent;
-import com.reactnativecommunity.webview.events.TopMessageEvent;
-import com.reactnativecommunity.webview.events.TopOpenWindowEvent;
-import com.reactnativecommunity.webview.events.TopRenderProcessGoneEvent;
-import com.reactnativecommunity.webview.events.TopShouldStartLoadWithRequestEvent;
-
-import java.util.Map;
-
-public class RNCWebViewManager extends ViewGroupManager<RNCWebViewWrapper> {
-
-    private final RNCWebViewManagerImpl mRNCWebViewManagerImpl;
-
-    public RNCWebViewManager() {
-        mRNCWebViewManagerImpl = new RNCWebViewManagerImpl();
-    }
-
-    @Override
-    public String getName() {
-        return RNCWebViewManagerImpl.NAME;
-    }
-
-    @Override
-    public RNCWebViewWrapper createViewInstance(ThemedReactContext context) {
-        return mRNCWebViewManagerImpl.createViewInstance(context);
-    }
-
-    public RNCWebViewWrapper createViewInstance(ThemedReactContext context, RNCWebView view) {
-      return mRNCWebViewManagerImpl.createViewInstance(context, view);
-    }
-
-    @ReactProp(name = "allowFileAccess")
-    public void setAllowFileAccess(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setAllowFileAccess(view, value);
-    }
-
-    @ReactProp(name = "allowFileAccessFromFileURLs")
-    public void setAllowFileAccessFromFileURLs(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setAllowFileAccessFromFileURLs(view, value);
-
-    }
-
-    @ReactProp(name = "allowUniversalAccessFromFileURLs")
-    public void setAllowUniversalAccessFromFileURLs(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setAllowUniversalAccessFromFileURLs(view, value);
-    }
-
-    @ReactProp(name = "allowsFullscreenVideo")
-    public void setAllowsFullscreenVideo(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setAllowsFullscreenVideo(view, value);
-    }
-
-    @ReactProp(name = "allowsProtectedMedia")
-    public void setAllowsProtectedMedia(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setAllowsProtectedMedia(view, value);
-    }
-
-    @ReactProp(name = "androidLayerType")
-    public void setAndroidLayerType(RNCWebViewWrapper view, @Nullable String value) {
-        mRNCWebViewManagerImpl.setAndroidLayerType(view, value);
-    }
-
-    @ReactProp(name = "applicationNameForUserAgent")
-    public void setApplicationNameForUserAgent(RNCWebViewWrapper view, @Nullable String value) {
-        mRNCWebViewManagerImpl.setApplicationNameForUserAgent(view, value);
-    }
-
-    @ReactProp(name = "basicAuthCredential")
-    public void setBasicAuthCredential(RNCWebViewWrapper view, @Nullable ReadableMap value) {
-        mRNCWebViewManagerImpl.setBasicAuthCredential(view, value);
-    }
-
-    @ReactProp(name = "cacheEnabled")
-    public void setCacheEnabled(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setCacheEnabled(view, value);
-    }
-
-    @ReactProp(name = "cacheMode")
-    public void setCacheMode(RNCWebViewWrapper view, @Nullable String value) {
-        mRNCWebViewManagerImpl.setCacheMode(view, value);
-    }
-
-    @ReactProp(name = "domStorageEnabled")
-    public void setDomStorageEnabled(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setDomStorageEnabled(view, value);
-    }
-
-    @ReactProp(name = "downloadingMessage")
-    public void setDownloadingMessage(RNCWebViewWrapper view, @Nullable String value) {
-        mRNCWebViewManagerImpl.setDownloadingMessage(value);
-    }
-
-    @ReactProp(name = "forceDarkOn")
-    public void setForceDarkOn(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setForceDarkOn(view, value);
-    }
-
-    @ReactProp(name = "geolocationEnabled")
-    public void setGeolocationEnabled(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setGeolocationEnabled(view, value);
-    }
-
-    @ReactProp(name = "hasOnScroll")
-    public void setHasOnScroll(RNCWebViewWrapper view, boolean hasScrollEvent) {
-        mRNCWebViewManagerImpl.setHasOnScroll(view, hasScrollEvent);
-    }
-
-    @ReactProp(name = "incognito")
-    public void setIncognito(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setIncognito(view, value);
-    }
-
-    @ReactProp(name = "injectedJavaScript")
-    public void setInjectedJavaScript(RNCWebViewWrapper view, @Nullable String value) {
-        mRNCWebViewManagerImpl.setInjectedJavaScript(view, value);
-    }
-
-    @ReactProp(name = "injectedJavaScriptBeforeContentLoaded")
-    public void setInjectedJavaScriptBeforeContentLoaded(RNCWebViewWrapper view, @Nullable String value) {
-        mRNCWebViewManagerImpl.setInjectedJavaScriptBeforeContentLoaded(view, value);
-    }
-
-    @ReactProp(name = "injectedJavaScriptForMainFrameOnly")
-    public void setInjectedJavaScriptForMainFrameOnly(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setInjectedJavaScriptForMainFrameOnly(view, value);
-
-    }
-
-    @ReactProp(name = "injectedJavaScriptBeforeContentLoadedForMainFrameOnly")
-    public void setInjectedJavaScriptBeforeContentLoadedForMainFrameOnly(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setInjectedJavaScriptBeforeContentLoadedForMainFrameOnly(view, value);
-
-    }
-
-    @ReactProp(name = "injectedJavaScriptObject")
-    public void setInjectedJavaScriptObject(RNCWebViewWrapper view, @Nullable String value) {
-        mRNCWebViewManagerImpl.setInjectedJavaScriptObject(view, value);
-    }
-
-    @ReactProp(name = "javaScriptCanOpenWindowsAutomatically")
-    public void setJavaScriptCanOpenWindowsAutomatically(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setJavaScriptCanOpenWindowsAutomatically(view, value);
-    }
-
-    @ReactProp(name = "javaScriptEnabled")
-    public void setJavaScriptEnabled(RNCWebViewWrapper view, boolean enabled) {
-        mRNCWebViewManagerImpl.setJavaScriptEnabled(view, enabled);
-    }
-
-    @ReactProp(name = "lackPermissionToDownloadMessage")
-    public void setLackPermissionToDownloadMessage(RNCWebViewWrapper view, @Nullable String value) {
-        mRNCWebViewManagerImpl.setLackPermissionToDownloadMessage(value);
-    }
-
-    @ReactProp(name = "hasOnOpenWindowEvent")
-    public void setHasOnOpenWindowEvent(RNCWebViewWrapper view, boolean hasEvent) {
-        mRNCWebViewManagerImpl.setHasOnOpenWindowEvent(view, hasEvent);
-    }
-
-    @ReactProp(name = "mediaPlaybackRequiresUserAction")
-    public void setMediaPlaybackRequiresUserAction(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setMediaPlaybackRequiresUserAction(view, value);
-    }
-
-    @ReactProp(name = "messagingEnabled")
-    public void setMessagingEnabled(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setMessagingEnabled(view, value);
-    }
-
-    @ReactProp(name = "menuItems")
-    public void setMenuCustomItems(RNCWebViewWrapper view, @Nullable ReadableArray items) {
-        mRNCWebViewManagerImpl.setMenuCustomItems(view, items);
-    }
-
-    @ReactProp(name = "messagingModuleName")
-    public void setMessagingModuleName(RNCWebViewWrapper view, @Nullable String value) {
-        mRNCWebViewManagerImpl.setMessagingModuleName(view, value);
-    }
-
-    @ReactProp(name = "minimumFontSize")
-    public void setMinimumFontSize(RNCWebViewWrapper view, int value) {
-        mRNCWebViewManagerImpl.setMinimumFontSize(view, value);
-    }
-
-    @ReactProp(name = "mixedContentMode")
-    public void setMixedContentMode(RNCWebViewWrapper view, @Nullable String value) {
-        mRNCWebViewManagerImpl.setMixedContentMode(view, value);
-    }
-
-    @ReactProp(name = "nestedScrollEnabled")
-    public void setNestedScrollEnabled(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setNestedScrollEnabled(view, value);
-    }
-
-    @ReactProp(name = "overScrollMode")
-    public void setOverScrollMode(RNCWebViewWrapper view, @Nullable String value) {
-        mRNCWebViewManagerImpl.setOverScrollMode(view, value);
-    }
-
-    @ReactProp(name = "saveFormDataDisabled")
-    public void setSaveFormDataDisabled(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setSaveFormDataDisabled(view, value);
-    }
-
-    @ReactProp(name = "scalesPageToFit")
-    public void setScalesPageToFit(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setScalesPageToFit(view, value);
-    }
-
-    @ReactProp(name = "setBuiltInZoomControls")
-    public void setSetBuiltInZoomControls(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setSetBuiltInZoomControls(view, value);
-    }
-
-    @ReactProp(name = "setDisplayZoomControls")
-    public void setSetDisplayZoomControls(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setSetDisplayZoomControls(view, value);
-    }
-
-    @ReactProp(name = "setSupportMultipleWindows")
-    public void setSetSupportMultipleWindows(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setSetSupportMultipleWindows(view, value);
-    }
-
-    @ReactProp(name = "showsHorizontalScrollIndicator")
-    public void setShowsHorizontalScrollIndicator(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setShowsHorizontalScrollIndicator(view, value);
-    }
-
-    @ReactProp(name = "showsVerticalScrollIndicator")
-    public void setShowsVerticalScrollIndicator(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setShowsVerticalScrollIndicator(view, value);
-    }
-
-    @ReactProp(name = "source")
-    public void setSource(RNCWebViewWrapper view, @Nullable ReadableMap value) {
-        mRNCWebViewManagerImpl.setSource(view, value);
-    }
-
-    @ReactProp(name = "textZoom")
-    public void setTextZoom(RNCWebViewWrapper view, int value) {
-        mRNCWebViewManagerImpl.setTextZoom(view, value);
-    }
-
-    @ReactProp(name = "thirdPartyCookiesEnabled")
-    public void setThirdPartyCookiesEnabled(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setThirdPartyCookiesEnabled(view, value);
-    }
-
-    @ReactProp(name = "webviewDebuggingEnabled")
-    public void setWebviewDebuggingEnabled(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setWebviewDebuggingEnabled(view, value);
-    }
-
-    @ReactProp(name = "userAgent")
-    public void setUserAgent(RNCWebViewWrapper view, @Nullable String value) {
-        mRNCWebViewManagerImpl.setUserAgent(view, value);
-    }
-
-    @ReactProp(name = "paymentRequestEnabled")
-    public void setPaymentRequestEnabled(RNCWebViewWrapper view, boolean value) {
-        mRNCWebViewManagerImpl.setPaymentRequestEnabled(view, value);
-    }
-
-    @Override
-    protected void addEventEmitters(@NonNull ThemedReactContext reactContext, RNCWebViewWrapper viewWrapper) {
-        // Do not register default touch emitter and let WebView implementation handle touches
-        viewWrapper.getWebView().setWebViewClient(new RNCWebViewClient());
-    }
-
-    @Override
-    public Map<String, Object> getExportedCustomDirectEventTypeConstants() {
-        Map<String, Object> export = super.getExportedCustomDirectEventTypeConstants();
-        if (export == null) {
-            export = MapBuilder.newHashMap();
-        }
-        // Default events but adding them here explicitly for clarity
-        export.put(TopLoadingStartEvent.EVENT_NAME, MapBuilder.of("registrationName", "onLoadingStart"));
-        export.put(TopLoadingFinishEvent.EVENT_NAME, MapBuilder.of("registrationName", "onLoadingFinish"));
-        export.put(TopLoadingErrorEvent.EVENT_NAME, MapBuilder.of("registrationName", "onLoadingError"));
-        export.put(SubResourceErrorEvent.EVENT_NAME, MapBuilder.of("registrationName", "onLoadingSubResourceError"));
-        export.put(TopMessageEvent.EVENT_NAME, MapBuilder.of("registrationName", "onMessage"));
-        // !Default events but adding them here explicitly for clarity
-
-        export.put(TopLoadingProgressEvent.EVENT_NAME, MapBuilder.of("registrationName", "onLoadingProgress"));
-        export.put(TopShouldStartLoadWithRequestEvent.EVENT_NAME, MapBuilder.of("registrationName", "onShouldStartLoadWithRequest"));
-        export.put(ScrollEventType.getJSEventName(ScrollEventType.SCROLL), MapBuilder.of("registrationName", "onScroll"));
-        export.put(TopHttpErrorEvent.EVENT_NAME, MapBuilder.of("registrationName", "onHttpError"));
-        export.put(TopRenderProcessGoneEvent.EVENT_NAME, MapBuilder.of("registrationName", "onRenderProcessGone"));
-        export.put(TopCustomMenuSelectionEvent.EVENT_NAME, MapBuilder.of("registrationName", "onCustomMenuSelection"));
-        export.put(TopOpenWindowEvent.EVENT_NAME, MapBuilder.of("registrationName", "onOpenWindow"));
-        return export;
-    }
-
-    @Override
-    public @Nullable
-    Map<String, Integer> getCommandsMap() {
-        return mRNCWebViewManagerImpl.getCommandsMap();
-    }
-
-    @Override
-    public void receiveCommand(@NonNull RNCWebViewWrapper reactWebView, String commandId, @Nullable ReadableArray args) {
-        mRNCWebViewManagerImpl.receiveCommand(reactWebView, commandId, args);
-        super.receiveCommand(reactWebView, commandId, args);
-    }
-
-    @Override
-    protected void onAfterUpdateTransaction(@NonNull RNCWebViewWrapper view) {
-        super.onAfterUpdateTransaction(view);
-        mRNCWebViewManagerImpl.onAfterUpdateTransaction(view);
-    }
-
-    @Override
-    public void onDropViewInstance(@NonNull RNCWebViewWrapper view) {
-        mRNCWebViewManagerImpl.onDropViewInstance(view);
-        super.onDropViewInstance(view);
-    }
-}
