+++ apple/RNCWebViewImpl.h
@@ -150,9 +150,7 @@ shouldStartLoadForRequest:(NSMutableDictionary<NSString *, id> *)request
 - (void)stopLoading;
 - (void)requestFocus;
 - (void)clearCache:(BOOL)includeDiskFiles;
-#ifdef RCT_NEW_ARCH_ENABLED
 - (void)destroyWebView;
-#endif
 #if !TARGET_OS_OSX
 - (void)addPullToRefreshControl;
 - (void)pullToRefresh:(UIRefreshControl *)refreshControl;
