--- apple/RNCWebViewManager.h
@@ -1,9 +1,0 @@
-#ifndef RNCWebViewManager_h
-#define RNCWebViewManager_h
-
-#import <React/RCTViewManager.h>
-
-@interface RNCWebViewManager : RCTViewManager
-@end
-
-#endif /* RNCWebViewManager_h */
