+++ apple/RNCWebViewModule.mm
@@ -2,9 +2,7 @@
 
 #import "RNCWebViewDecisionManager.h"
 
-#ifdef RCT_NEW_ARCH_ENABLED
 #import <React/RCTFabricComponentsPlugins.h>
-#endif /* RCT_NEW_ARCH_ENABLED */
 
 @implementation RNCWebViewModule
 
@@ -21,11 +19,9 @@ RCT_EXPORT_METHOD(shouldStartLoadWithLockIdentifier:(BOOL)shouldStart lockIdenti
     [[RNCWebViewDecisionManager getInstance] setResult:shouldStart forLockIdentifier:(int)lockIdentifier];
 }
 
-#ifdef RCT_NEW_ARCH_ENABLED
 - (std::shared_ptr<facebook::react::TurboModule>)getTurboModule:(const facebook::react::ObjCTurboModule::InitParams &)params {
   return std::make_shared<facebook::react::NativeRNCWebViewModuleSpecJSI>(params);
 }
-#endif /* RCT_NEW_ARCH_ENABLED */
 
 Class RNCWebViewModuleCls(void) {
   return RNCWebViewModule.class;
