+++ example/macos/Podfile
@@ -2,4 +2,6 @@ require_relative '../../node_modules/react-native-test-app/macos/test_app'
 
 workspace 'WebviewExample.xcworkspace'
 
-use_test_app! :hermes_enabled => true
+# :new_arch_enabled is required until react-native-macos reaches 0.82+,
+# where the new architecture becomes the only option.
+use_test_app! :hermes_enabled => true, :new_arch_enabled => true
