+++ src/WebView.android.tsx
@@ -273,9 +273,6 @@ const WebViewComponent = forwardRef<unknown, AndroidWebViewProps>(
         hasOnOpenWindowEvent={onOpenWindowProp !== undefined}
         onShouldStartLoadWithRequest={onShouldStartLoadWithRequest}
         ref={webViewRef}
-        // TODO: find a better way to type this.
-        // @ts-expect-error source is old arch
-        source={sourceResolved}
         newSource={newSource}
         style={webViewStyles}
         overScrollMode={overScrollMode}
