+++ src/WebView.ios.tsx
@@ -240,8 +240,6 @@ const WebViewComponent = forwardRef<unknown, IOSWebViewProps>(
         style={webViewStyles}
         hasOnFileDownload={!!onFileDownload}
         ref={webViewRef}
-        // @ts-expect-error old arch only
-        source={sourceResolved}
         {...nativeConfig?.props}
       />
     );
